refactor: organize backend routes
